Fixing double enhance bug
The previous implementation would edit the message to enhance as the
message was being scanned for enhanceable text. This ended up resulting
in some strings being double enhanced because the replacement text
matched another enhancer's regex.

To handle this case, Enhancers now return Substitution objects, which
are applied after all Enhancers have had a chance to improve the
original message.